Move TriggerBaseEntity
